Refactor Database architecture and Tools organization
- Decouple `VirtualRelvarDefinition` from `Database` using new `QueryExecutor` trait
- Move `relvar-core/src/database.rs` to module structure `relvar-core/src/database/`
- Consolidate `exporter`, `importer`, `visualizer` into `relvar/src/tools/`
- Deprecate `relvar::data` and `relvar::visualizer` in favor of `relvar::tools`
