--- relvar/src/data/mod.rs
@@ -1,6 +1,0 @@
-//! Data import and export functionality.
-//!
-//! This module provides tools for moving data in and out of Relvar relations.
-
-pub mod exporter;
-pub mod importer;
